test: merge smaller test suites into one
